pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tigo-energy-ha"
version = "0.1.0"
description = "Home Assistant custom integration for Tigo Energy"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Drew Cotten" }]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
]
dependencies = []

[project.optional-dependencies]
dev = [
  "aioresponses>=0.7.8",
  "pytest>=8.2.0",
  "pytest-homeassistant-custom-component>=0.13.170",
  "ruff>=0.11.0"
]

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
